feat(ffmpeg): software fallback in FfmpegVideoStreamDecoder
The decoder now transparently falls back from the HW probe
(VideoToolbox / VAAPI / NVDEC / D3D11VA) to a software
`ffmpeg::decoder::Video` when AllBackendsFailed surfaces:

- At `open` time (no HW backend opens): SW decoder opens immediately
  from the same `Parameters`. `open` only returns Err when both paths
  fail.
- Mid-stream from `send_packet` / `receive_frame` / `send_eof`:
  the rescued `unconsumed_packets` from the HW probe are replayed
  through the freshly-opened SW decoder before the new packet (or the
  next receive_frame poll) is processed. Non-seekable inputs survive
  HW exhaustion without losing any compressed data.
- After the transition the decoder stays on SW (no probe-back-to-HW).

Internal state is a `Hw(VideoDecoder) | Sw(ffmpeg::decoder::Video)`
enum. Two scratch frames (one per path) avoid per-call allocations.
Public `is_software()` / `is_hardware()` / `hardware_inner()`
accessors let tests and tracing introspect.

Other pieces this required:

1. **Packet conversion helper** in boundary.rs:
   `ffmpeg_packet_from_video_packet(&VideoPacket<VideoPacketExtra,
   FfmpegBuffer>) -> ffmpeg::Packet` — copies the compressed bytes,
   forwards pts/dts/duration/flags/stream_index. Zero-copy passthrough
   of the FfmpegBuffer's AVBufferRef is a future optimization. Note
   that ffmpeg-next 8.x doesn't expose AV_PKT_FLAG_DISCARD on its
   `packet::Flags` API; we forward KEY + CORRUPT and silently drop
   DISCARD.

2. **Pixel-format table extensions in frame.rs**: SW decoders produce
   formats outside the original HW download set (NV* / P0xx / P2xx /
   P4xx). Added is_supported_cpu_pix_fmt + plane_height_for +
   plane_row_bytes_for arms for:
   - planar YUV 4:2:0 / 4:2:2 / 4:4:4 at 8 / 10 / 12 / 16-bit
   - packed RGB24 / BGR24 / RGBA / BGRA / ARGB / ABGR
   - GRAY8 / GRAY16LE
   The existing test that asserted these formats were "unsupported"
   was updated — they're now intentionally supported because of this
   commit. (YUVJ420P stays unsupported; the range info now lives on
   `ColorInfo::range`, so the J variants don't get a separate
   identifier.)

`send_packet` is now fully wired for both HW and SW paths
(was returning SendPacketNotImplemented before). The
`SendPacketNotImplemented` error variant is gone.

127 workspace tests pass. cargo build --workspace clean.

+++ mediadecode-ffmpeg/src/buffer.rs
@@ -88,7 +88,7 @@ impl FfmpegBuffer {
     if buf.is_null() {
       return None;
     }
-    let buf_size = unsafe { (*buf).size as usize };
+    let buf_size = unsafe { (*buf).size };
     let end = offset.checked_add(len)?;
     if end > buf_size {
       return None;
@@ -121,7 +121,7 @@ impl FfmpegBuffer {
     if buf.is_null() {
       return None;
     }
-    let len = unsafe { (*buf).size as usize };
+    let len = unsafe { (*buf).size };
     Some(Self {
       inner: buf,
       offset: 0,
